+++ components/forms/input/index.jsx
@@ -17,6 +17,11 @@ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
 // React is an external dependency of the project.
 import React from 'react';
 
+// ### shortid
+// [npmjs.com/package/shortid](https://www.npmjs.com/package/shortid)
+// shortid is a short, non-sequential, url-friendly, unique id generator
+import shortid from 'shortid';
+
 // ### classNames
 // [github.com/JedWatson/classnames](https://github.com/JedWatson/classnames)
 // This project uses `classnames`, "a simple javascript utility for conditionally
@@ -27,6 +32,11 @@ import classNames from 'classnames';
 import InputIcon from '../../icon/input-icon';
 // This component's `checkProps` which issues warnings to developers about properties when in development mode (similar to React's built in development tools)
 import checkProps from './check-props';
+
+// ### isFunction
+import isFunction from 'lodash.isfunction';
+
+import Button from '../../button';
 
 // Remove the need for `React.PropTypes`
 const { PropTypes } = React;
@@ -97,9 +107,13 @@ const Input = React.createClass({
 			'right'
 		]),
 		/**
+		 * Set the assistive text for a clickable icon
+		 */
+		iconAssistiveText: PropTypes.string,
+		/**
 		 * Every input must have a unique ID in order to support keyboard navigation and ARIA support.
 		 */
-		id: PropTypes.string.isRequired,
+		id: PropTypes.string,
 		/**
 		 * This label appears above the input.
 		 */
@@ -108,6 +122,10 @@ const Input = React.createClass({
 		 * This event fires when the input changes.
 		 */
 		onChange: PropTypes.func,
+		/**
+		 * This event fires when the input is clicked.
+		 */
+		onClick: PropTypes.func,
 		/**
 		 * This event fires when the icon is clicked.
 		 */
@@ -152,6 +170,8 @@ const Input = React.createClass({
 	componentWillMount () {
 		// `checkProps` issues warnings to developers about properties (similar to React's built in development tools)
 		checkProps(FORMS_INPUT, this.props);
+
+		this.generatedId = shortid.generate();
 	},
 
 	getDefaultProps () {
@@ -159,6 +179,29 @@ const Input = React.createClass({
 			iconPosition: 'left',
 			type: 'text'
 		};
+	},
+
+	getId () {
+		return this.props.id || this.generatedId;
+	},
+
+	getIconRender (position) {
+		if (position !== this.props.iconPosition) return '';
+
+		return isFunction(this.props.onIconClick)
+		? (<Button
+			iconVariant="small"
+			variant="icon"
+			className="slds-input__icon slds-button--icon"
+			assistiveText={this.props.iconAssistiveText}
+			iconName={this.props.iconName}
+			iconCategory={this.props.iconCategory}
+			onClick={this.props.onIconClick}
+		/>)
+		: (<InputIcon
+			name={this.props.iconName}
+			category={this.props.iconCategory}
+		/>);
 	},
 
 	// ### Render
@@ -173,10 +216,11 @@ const Input = React.createClass({
 			iconCategory,
 			iconName,
 			iconPosition,
-			id,
+			inlineEditTrigger, // eslint-disable-line react/prop-types
 			label,
 			onChange,
-			onIconClick,
+			onClick,
+			onIconClick, // eslint-disable-line no-unused-vars
 			placeholder,
 			readOnly,
 			required,
@@ -191,7 +235,7 @@ const Input = React.createClass({
 		const hasIcon = iconCategory && iconName;
 
 		// One of these is required to pass accessibility tests
-		const labelText = assistiveText || label;
+		const labelText = label || assistiveText;
 
 		return (
 			<div
@@ -201,42 +245,49 @@ const Input = React.createClass({
 				},
 				className)}
 			>
-				{labelText && (!readOnly
-					? <label className={classNames('slds-form-element__label', { 'slds-assistive-text': assistiveText })} htmlFor={id}>
+				{labelText && (readOnly
+					? <span
+						className={classNames('slds-form-element__label', { 'slds-assistive-text': assistiveText && !label })}
+					>
+						{labelText}
+					</span>
+					: <label
+						className={classNames('slds-form-element__label', { 'slds-assistive-text': assistiveText && !label })}
+						htmlFor={this.getId()}
+					>
 						{required && <abbr className="slds-required" title="required">*</abbr>}
 						{labelText}
 					</label>
-					: <span className={classNames('slds-form-element__label', { 'slds-assistive-text': assistiveText })}>
-						{labelText}
-					</span>)}
+				)}
 				<div
 					className={classNames('slds-form-element__control', hasIcon && [
 						'slds-input-has-icon',
 						`slds-input-has-icon--${iconPosition}`
 					], {
-						'slds-has-divider--bottom': readOnly
+						'slds-has-divider--bottom': readOnly && !inlineEditTrigger
 					})}
 				>
-					{hasIcon && <InputIcon
-						name={iconName}
-						category={iconCategory}
-						onClick={onIconClick}
-					/>}
+					{hasIcon && this.getIconRender('left')}
+
 					{!readOnly && <input
 						{...props}
 						aria-controls={ariaControls}
 						aria-owns={ariaOwns}
 						className="slds-input"
 						disabled={disabled}
-						id={id}
+						id={this.getId()}
 						onChange={onChange}
+						onClick={onClick}
 						placeholder={placeholder}
 						required={required}
 						type={type}
 						value={value}
 					/>}
-					{readOnly && <span className="slds-form-element__static">
+					{hasIcon && this.getIconRender('right')}
+
+					{readOnly && <span className="slds-form-element__static" onClick={onClick}>
 						{value}
+						{inlineEditTrigger}
 					</span>}
 				</div>
 				{errorText && <div className="slds-form-element__help">{errorText}</div>}