+++ stories/index.js
@@ -2,6 +2,13 @@ import '@salesforce-ux/design-system/assets/styles/salesforce-lightning-design-s
 
 export default from './bread-crumb';
 export default from './button';
+export default from './card';
 export default from './context-bar';
 export default from './data-table';
+export default from './date-picker';
+export default from './forms/input';
+export default from './forms/input/inline';
 export default from './icon';
+export default from './lookup';
+export default from './spinner';
+export default from './time-picker';